refactor step files
(cherry picked from commit 7d22cec)
